+++ out-adapter-persistence-jpa/src/main/java/gohigher/user/UserPersistenceQueryAdapter.java
@@ -4,6 +4,8 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
+import gohigher.user.entity.UserJpaEntity;
+import gohigher.user.entity.UserRepository;
 import gohigher.user.port.out.UserPersistenceQueryPort;
 import lombok.RequiredArgsConstructor;
